pyproject.toml
[tool.poetry]
name = "simple-python-api-demo"
version = "0.1.0"
description = ""
authors = ["Lucas Roe <lroe2930@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-dotenv = "^1.0.0"
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
